refactor(p2p): Split large files into focused modules (Waves 1-3)
Split 7 large files (~8k lines) into 38 smaller focused modules:

- message.rs (1,350 lines) → message/ (7 files)
- bitswap/access.rs (544 lines) → access.rs + registry.rs
- host.rs (2,059 lines) → host/ (6 files)
- sync/peer_state.rs (957 lines) → sync/peer_state/ (3 files)
- sync/dag_sync.rs (588 lines) → sync/dag_sync/ (5 files)
- sync/manager.rs (956 lines) → sync/manager/ (5 files)
- two_stream.rs (790 lines) → two_stream/ (4 files)

All public API exports preserved at same paths. Wire format unchanged.
176 tests pass. Clippy clean.
