+++ crates/p2p/src/message/metadata.rs
@@ -1,0 +1,59 @@
+//! Message metadata for P2P protocol messages.
+
+use serde::{Deserialize, Serialize};
+
+use super::cbor::{nullable_bytes, optional_bytes};
+use crate::protocol::MESSAGE_VERSION;
+
+/// Metadata that is part of every P2P message.
+///
+/// This struct contains common fields for message routing, authentication,
+/// and error handling.
+#[derive(Debug, Clone, Serialize, Deserialize, Default)]
+pub struct MetaData {
+    /// DefraDB message version.
+    #[serde(rename = "Version")]
+    pub version: String,
+
+    /// Unique message identifier. Responses use the same ID as the request.
+    #[serde(rename = "MessageID")]
+    pub message_id: String,
+
+    /// ID of the sender (PeerID when using libp2p).
+    #[serde(rename = "SenderID")]
+    pub sender_id: String,
+
+    /// Public key of the node that created the message.
+    /// Uses nullable_bytes to handle Go's nil []byte as CBOR null.
+    #[serde(rename = "Pubkey", with = "nullable_bytes")]
+    pub pubkey: Vec<u8>,
+
+    /// Signature for message authentication.
+    /// Uses custom serialization for optional CBOR bytes.
+    #[serde(
+        rename = "Signature",
+        skip_serializing_if = "Option::is_none",
+        default,
+        with = "optional_bytes"
+    )]
+    pub signature: Option<Vec<u8>>,
+
+    /// Error message if something went wrong.
+    #[serde(rename = "ErrMessage", skip_serializing_if = "Option::is_none")]
+    pub err_message: Option<String>,
+}
+
+impl MetaData {
+    /// Create new metadata with default values.
+    pub fn new() -> Self {
+        Self {
+            version: MESSAGE_VERSION.to_string(),
+            ..Default::default()
+        }
+    }
+
+    /// Set the message version to the current protocol version.
+    pub fn set_version(&mut self) {
+        self.version = MESSAGE_VERSION.to_string();
+    }
+}
